move schema to v1::schema
