+++ rust/parquet/src/arrow/record_reader.rs
@@ -21,7 +21,7 @@ use std::mem::{replace, size_of};
 use crate::column::{page::PageReader, reader::ColumnReaderImpl};
 use crate::data_type::DataType;
 use crate::errors::{ParquetError, Result};
-use crate::schema::types::ColumnDescPtr;
+use crate::schema::v1::types::ColumnDescPtr;
 use arrow::array::BooleanBufferBuilder;
 use arrow::bitmap::Bitmap;
 use arrow::buffer::{Buffer, MutableBuffer};
@@ -439,8 +439,8 @@ mod tests {
     use crate::column::page::PageReader;
     use crate::data_type::Int32Type;
     use crate::errors::Result;
-    use crate::schema::parser::parse_message_type;
-    use crate::schema::types::SchemaDescriptor;
+    use crate::schema::v1::parser::parse_message_type;
+    use crate::schema::v1::types::SchemaDescriptor;
     use crate::util::test_common::page_util::{DataPageBuilder, DataPageBuilderImpl};
     use arrow::array::{BooleanBufferBuilder, Int16BufferBuilder, Int32BufferBuilder};
     use arrow::bitmap::Bitmap;
