+++ rust/parquet/src/column/mod.rs
@@ -45,7 +45,7 @@
 //!         reader::{FileReader, SerializedFileReader},
 //!         writer::{FileWriter, SerializedFileWriter},
 //!     },
-//!     schema::parser::parse_message_type,
+//!     schema::v1::parser::parse_message_type,
 //! };
 //!
 //! let path = Path::new("/path/to/column_sample.parquet");
